+++ app/src/main/java/gitmad/gitmadheatmap/ILocationCallback.java
@@ -7,6 +7,6 @@ import java.util.List;
 /**
  * Simple callback interface.
  */
-public interface LocationCallback {
+public interface ILocationCallback {
     void onFinish(List<LatLng> locations);
 }
