+++ app/src/main/java/gitmad/gitmadheatmap/LoginActivity.java
@@ -83,7 +83,7 @@ public class LoginActivity extends AppCompatActivity {
      * is already signed in and opening the app.
      */
     private void transitionToEnterActivity() {
-        Intent intent = new Intent(this, UserLoggedInActivity.class);
+        Intent intent = new Intent(this, MainActivity.class);
         intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
         intent.putExtra(Integer.toString(R.string.intent_menu_item), "nav_home_option");
         startActivity(intent);
