+++ app/src/main/res/layout/activity_maps.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:orientation="vertical"
-        tools:context=".UserLoggedInActivity">
+        tools:context=".MainActivity">
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/loggedIn_toolbar"
