+++ app/src/main/res/layout/fragment_enter.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EnterFragment">
+    tools:context=".AlarmFragment">
 
     <Button
         android:id="@+id/enter_btn_start_alarm"
